implement alert

+++ src/app/pages/home/home.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { ProductsService } from 'src/app/services/products/products.service';
+import { ESizeModal } from 'src/app/shared/utils/modal-size.enum';
 import { IDataRecord } from 'src/app/shared/utils/records.interface';
 import { ETypesButton } from 'src/app/shared/utils/type-button.enum';
+import { AlertService } from '../../services/alert/alert.service';
+import { EAlertType } from 'src/app/shared/utils/alert-type.enum';
 
 @Component({
   selector: 'app-home',
@@ -17,8 +20,14 @@ export class HomeComponent implements OnInit {
   typeButton = ETypesButton;
   showData: IDataRecord[] = [];
   searchTerm = '';
+  itemSelected: IDataRecord | null = null;
+  showModalConfirm = false;
+  sizeModal = ESizeModal;
 
-  constructor(private productsService: ProductsService) {}
+  constructor(
+    private productsService: ProductsService,
+    private alertService: AlertService,
+  ) {}
 
   ngOnInit() {
     /**
@@ -41,6 +50,10 @@ export class HomeComponent implements OnInit {
         //this.productsService.removeAllProducts(data);
       },
       error: (error) => {
+        this.alertService.message$.next({
+          description: 'Ocurrió un error al cargar los productos',
+          type: EAlertType.ERROR,
+        });
         console.error('Error', error);
       },
       complete: () => {
@@ -76,20 +89,35 @@ export class HomeComponent implements OnInit {
     this.showData = filteredData.slice(initalIndex, finalIndex);
   }
 
-  deleteProduct(item: IDataRecord) {
-    console.clear();
-    this.productsService.deleteProduct(item.id).subscribe({
-      next: (data) => {
-        console.log('aca', data);
-        this.loadProducts();
-      },
-      error: (error) => {
-        console.error('Error', error);
-      },
-      complete: () => {
-        console.log('Complete');
-      },
-    });
+  selectItem(item: IDataRecord) {
+    this.itemSelected = item;
+    this.showModalConfirm = true;
+  }
+
+  deleteProduct() {
+    if (this.itemSelected) {
+      const item = { ...this.itemSelected };
+      this.productsService.deleteProduct(item.id).subscribe({
+        next: () => {
+          this.alertService.message$.next({
+            description: `Producto: ${item.name} eliminado`,
+            type: EAlertType.SUCCESS,
+          });
+          this.loadProducts();
+        },
+        error: (error) => {
+          this.alertService.message$.next({
+            description: `Ocurrió un error al eliminar el producto: ${item.name}`,
+            type: EAlertType.ERROR,
+          });
+          console.error('Error', error);
+        },
+        complete: () => {
+          this.itemSelected = null;
+          this.showModalConfirm = false;
+        },
+      });
+    }
   }
 
   editProduct(item: IDataRecord) {
